fix: use i64 for shards/ranges
Using int is problematic because it's platform-dependent. On 32-bit
platform this may cause issues if SubjectShards > 2,147,483,647. This
will make the number go negative and ranges might not match.

+++ eppoclient/client.go
@@ -89,12 +89,12 @@ func (ec *EppoClient) getAssignment(subjectKey string, flagKey string, subjectAt
 
 	// Get assigned variation
 	assignmentKey := "assignment-" + subjectKey + "-" + flagKey
-	shard := getShard(assignmentKey, int64(config.SubjectShards))
+	shard := getShard(assignmentKey, config.SubjectShards)
 	variations := allocation.Variations
 	var variationShard Variation
 
 	for _, variation := range variations {
-		if isShardInRange(int(shard), variation.ShardRange) {
+		if isShardInRange(shard, variation.ShardRange) {
 			variationShard = variation
 		}
 	}
@@ -137,9 +137,9 @@ func getSubjectVariationOverride(experimentConfig experimentConfiguration, subje
 	return Null()
 }
 
-func isInExperimentSample(subjectKey string, flagKey string, subjectShards int, percentExposure float32) bool {
+func isInExperimentSample(subjectKey string, flagKey string, subjectShards int64, percentExposure float32) bool {
 	shardKey := "exposure-" + subjectKey + "-" + flagKey
-	shard := getShard(shardKey, int64(subjectShards))
+	shard := getShard(shardKey, subjectShards)
 
 	return float64(shard) <= float64(percentExposure)*float64(subjectShards)
 }
